requirements.txt
torchvision
torchtext==0.6.0
pyobjc-framework-Quartz
pyminitouch
pynput
